+++ device-bootloaders/ServiceProcessorBL2/services/historical_extreme.c
@@ -53,14 +53,14 @@ static void get_max_memory_error(tag_id_t tag_id, uint64_t req_start_time)
         if (error_count->ddr_ce_max_count < ddr_ce_count) {
             error_count-> ddr_ce_max_count = ddr_ce_count;
         }
-        // TODO: Change this to uint32_t in rsp control block. same value is present in the json schema as well sw-6467
-        dm_rsp.max_ecc_count.count = (uint8_t)error_count->ddr_ce_max_count; 
+
+        dm_rsp.max_ecc_count.count = error_count->ddr_ce_max_count; 
     }
 
     FILL_RSP_HEADER(dm_rsp, tag_id,
                     DM_CMD_GET_MAX_MEMORY_ERROR,
                     timer_get_ticks_count() - req_start_time,
-                    (uint32_t)status); // TODO update the status type - SW-6467
+                    status);
  
 
     if (0 != SP_Host_Iface_CQ_Push_Cmd((char *)&dm_rsp, sizeof(struct device_mgmt_max_memory_error_rsp_t))) {
@@ -92,7 +92,7 @@ static void get_module_max_ddr_bw(tag_id_t tag_id, uint64_t req_start_time)
 {
     struct device_mgmt_max_dram_bw_rsp_t dm_rsp;
     struct max_dram_bw_t max_dram_bw;
-    int status;
+    int32_t status;
 
     status = get_module_max_dram_bw(&max_dram_bw);
 
@@ -106,7 +106,7 @@ static void get_module_max_ddr_bw(tag_id_t tag_id, uint64_t req_start_time)
     FILL_RSP_HEADER(dm_rsp, tag_id,
                     DM_CMD_GET_MODULE_MAX_DDR_BW,
                     timer_get_ticks_count() - req_start_time,
-                    (uint32_t)status);
+                    status);
 
     if (0 != SP_Host_Iface_CQ_Push_Cmd((char *)&dm_rsp, sizeof(struct device_mgmt_max_dram_bw_rsp_t))) {
         printf("get_max_memory_error: Cqueue push error!\n");
@@ -138,7 +138,7 @@ static void get_module_max_throttle_time(tag_id_t tag_id, uint64_t req_start_tim
 {
     struct device_mgmt_max_throttle_time_rsp_t dm_rsp;
     uint64_t max_throttle_time;
-    int status;
+    int32_t status;
 
     status = get_max_throttle_time(&max_throttle_time);
 
@@ -151,7 +151,7 @@ static void get_module_max_throttle_time(tag_id_t tag_id, uint64_t req_start_tim
     FILL_RSP_HEADER(dm_rsp, tag_id,
                     DM_CMD_GET_MODULE_MAX_THROTTLE_TIME,
                     timer_get_ticks_count() - req_start_time,
-                    (uint32_t)status);
+                    status);
 
     if (0 != SP_Host_Iface_CQ_Push_Cmd((char *)&dm_rsp, sizeof(struct device_mgmt_max_throttle_time_rsp_t))) {
         printf("get_module_max_throttle_time: Cqueue push error!\n");
@@ -184,7 +184,7 @@ static void get_module_max_temperature(uint16_t tag, uint64_t req_start_time)
 {
     struct device_mgmt_max_temperature_rsp_t dm_rsp;
     uint8_t max_temp;
-    int status;
+    int32_t status;
 
     status = get_soc_max_temperature(&max_temp);
 
@@ -197,7 +197,7 @@ static void get_module_max_temperature(uint16_t tag, uint64_t req_start_time)
     FILL_RSP_HEADER(dm_rsp, tag,
                     DM_CMD_GET_MODULE_MAX_TEMPERATURE,
                     timer_get_ticks_count() - req_start_time,
-                    (uint32_t)status);
+                    status);
 
 
    if (0 != SP_Host_Iface_CQ_Push_Cmd((char *)&dm_rsp, sizeof(struct device_mgmt_max_temperature_rsp_t))) {
